packages/graphrag/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphrag"
version = "0.1.0"
description = "Neo4j GraphRAG (retrieval + LLM ranking) for protein-engineering"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "neo4j>=5.14.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
hf_local = [
  "transformers>=4.40.0",
  "torch",
]

[project.scripts]
graphrag = "graphrag.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
